+++ back/src/main/java/im2agoracle/univgrenoblealpes/fr/gromed/keys/AvisSMRKey.java
@@ -8,16 +8,16 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class AvisSMRKey implements Serializable {
 
-    @Column(name = "CODECISMEDICAMENT")
-    private int codeCIS_Medicament;
+    @Column(name="CODECISMEDICAMENT")
+    private long codeCIS_Medicament;
 
-    @Column(name = "CODEHASPAGE")
+    @Column(name="CODEHASPAGE")
     private int codeHAS_PageCT;
 
     private AvisSMRKey() {
     }
 
-    public int getCodeCIS() {
+    public long getCodeCIS() {
         return this.codeCIS_Medicament;
     }
